Move domain model user to `domain/model`
