+++ src/booking/domain/model/user/user.ts
@@ -1,4 +1,4 @@
-import { Entity, Identifier } from '../shared';
+import { Entity, Identifier } from '../../shared';
 
 export class UserId extends Identifier<string> {}
