+++ src/booking/usecase/user/__tests__/register.spec.ts
@@ -1,4 +1,4 @@
-import { User } from '../../../domain/user/user';
+import { User } from '../../../domain/model/user/user';
 import IUserRepository from '../i-repository';
 import { MemeryUserRepository } from '../../../adapter/repository/user/repository';
 import {
